+++ src/Mc.PowerShell/Mc.psm1
@@ -136,6 +136,7 @@ Export-ModuleMember -Function @(
     'Convert-McMarkdown'
     'Get-McSegmentSlice'
     'Show-McDriveChooser'
+    'Get-McDriveLocation'
     'Show-McSortMenu'
     'Get-McPanelCurrent'
     'Move-McPanelCursor'